backend/requirements.txt
fastapi[standard]
uvicorn[standard]
httpx
python-dotenv
python-docx
yt-dlp
